feat: refactor KGraph schema and routes, remove deprecated kgraph.js
TODO: 중복된 kgraph 수정, 수정된 kgraphSchema, messageSchema에 맞게 route 수정 필요
