+++ api/models/schema/messageSchema.js
@@ -1,5 +1,6 @@
 const mongoose = require('mongoose');
 const mongoMeili = require('~/models/plugins/mongoMeili');
+const { nodeSchema } = require('./kgraphSchema-removal');
 const messageSchema = mongoose.Schema(
   {
     messageId: {
@@ -54,21 +55,20 @@ const messageSchema = mongoose.Schema(
       type: String,
       meiliIndex: true,
     },
-    tempNodes: { 
-  type: [mongoose.Schema.Types.Mixed],
-  default: undefined,
-},
-nodes: {
-      type: Array, // 또는 [nodeSchema] (만약 노드 스키마를 정의했다면)
+    nodes: {
+      type: [{
+        nodeSchema,
+        isCurated: {
+          type: Boolean,
+          default : false,
+        },
+      }], // 또는 [nodeSchema] (만약 노드 스키마를 정의했다면)
       default: [],
     },
-isImported: { 
-  type: Boolean,
-  default: false,
-},
-isCreatedByUser: {
-  type: Boolean,
-},
+    isImported: {
+      type: Boolean,
+      default: false,
+    },
     summary: {
       type: String,
     },
@@ -153,16 +153,6 @@ isCreatedByUser: {
       default: undefined,
     },
     */
-    temp_nodes: {
-      type: [
-        {
-          nodeSummary: { type: String },
-          isCurated: { type: Boolean, default: false },
-        },
-      ],
-      default: undefined,
-    },
-
   },
   { timestamps: true },
 );
@@ -175,8 +165,6 @@ if (process.env.MEILI_HOST && process.env.MEILI_MASTER_KEY) {
     primaryKey: 'messageId',
   });
 }
-
-
 
 messageSchema.index({ createdAt: 1 });
 messageSchema.index({ messageId: 1, user: 1 }, { unique: true });
