+++ api/server/routes/index.js
@@ -26,7 +26,6 @@ const user = require('./user');
 const ask = require('./ask');
 const banner = require('./banner');
 const kgraphs = require('./kgraphs');
-const kgraphRouter = require('./kgraph');
 
 module.exports = {
   ask,
@@ -57,5 +56,4 @@ module.exports = {
   staticRoute,
   banner,
   kgraphs,
-  kgraphRouter,
 };
